move search, merge out of mdrepo and remove entity_set altogether
